requirements.txt
# Core web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.17
pgvector>=0.2.4
asyncpg>=0.29.0
alembic>=1.13.1

# Data validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# AI/ML
sentence-transformers>=2.3.1
huggingface-hub>=0.20.2
# torch>=2.1.2  # Commented out temporarily - enable Long Path support in Windows first
transformers>=4.36.2

# Skill extraction
rapidfuzz>=3.6.1
regex>=2023.12.25

# Vietnamese NLP (optional)
underthesea>=6.8.0

# File parsing
pypdf>=3.17.4
pdfplumber>=0.10.3
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.2.0

# OCR
pytesseract>=0.3.10
Pillow>=10.2.0
pdf2image>=1.17.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.3
